Add an ADR and move git code to pkg
