+++ src/pkg/git/auth_test.go
